richmenu/requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
frozenlist==1.4.1
future==1.0.0
idna==3.7
line-bot-sdk==3.10.1
linebot==0.1.6
loguru==0.7.2
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
